+++ src/KnoqReminder.App/KnoqReminder.App/Configurations/ServiceCollectionExtension.cs
@@ -8,8 +8,6 @@ static class ServiceCollectionExtension
     public static IServiceCollection ConfigureAppOptions(this IServiceCollection services, IConfigurationRoot config)
     {
         return services
-            .Configure<IDbConnectionOptions, MariaDbConnectionConfiguration>(config)
-            .Configure<IDefaultLocalizationOptions, DefaultLocalizationConfiguration>(config.GetSection(DefaultLocalizationConfiguration.Position))
             .Configure<IKnoqApiClientOptions, KnoqApiClientConfiguration>(config)
             .Configure<ITraqApiClientOptions, TraqApiClientConfiguration>(config)
             .Configure<ITraqBotOptions, TraqBotConfiguration>(config);
